pyproject.toml
[project]
name = "cardlens-research-terminal"
version = "0.1.0"
description = ""
authors = [
    {name = "YuanTeng",email = "yuanteng.fan1208@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.54.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pandas (>=1.4,<3)",
    "numpy (>=2.4.2,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "yfinance (>=1.2.0,<2.0.0)",
    "ta (>=0.11.0,<0.12.0)",
    "pdfplumber (>=0.11.9,<0.12.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "openai (>=2.21.0,<3.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "plotly (>=6.5.2,<7.0.0)",
    "feedparser (>=6.0.12,<7.0.0)",
    "lxml (>=6.0.2,<7.0.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.1,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)"
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
# E402: module-level import not at top — required for sys.path setup in Streamlit pages
# E501: line length handled by line-length setting
# E702: multiple statements on one line — accepted in compact signal logic
ignore = ["E402", "E501", "E702"]

[tool.pytest.ini_options]
testpaths = ["tests"]
